challenge-03/peguerosdc/requirements-dwave.txt
absl-py==0.11.0
anyio==2.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.11
dimod==0.9.13
dwave-cloud-client==0.8.4
dwave-greedy==0.1.2
dwave-hybrid==0.6.1
dwave-inspector==0.2.5
dwave-inspectorapp==0.2.2
dwave-neal==0.5.7
dwave-networkx==0.8.8
dwave-ocean-sdk==3.3.0
dwave-qbsolv==0.3.2
dwave-system==1.4.0
dwave-tabu==0.3.1
dwavebinarycsp==0.1.2
entrypoints==0.3
fasteners==0.16
Flask==1.1.2
funcsigs==1.0.2
homebase==1.0.1
idna==2.10
immutables==0.6
importlib-resources==5.1.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyterlab==3.0.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.1
kiwisolver==1.3.1
lark==0.11.1
MarkupSafe==1.1.1
matplotlib==3.3.3
minorminer==0.2.5
mistune==0.8.4
mock==4.0.3
mpmath==1.1.0
msgpack==0.6.2
nbclassic==0.2.6
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
notebook-as-pdf==0.4.0
numpy==1.19.5
ortools==8.1.8487
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
penaltymodel==0.16.4
penaltymodel-cache==0.4.3
penaltymodel-lp==0.1.4
penaltymodel-mip==0.2.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plucky==0.4.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
pycparser==2.20
pyee==8.1.0
Pygments==2.8.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.5
pyqubo==1.0.10
pyquil==2.27.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-rapidjson==1.0
pytz==2021.1
pyzmq==21.0.1
quantum-grove==1.7.0
requests==2.25.1
rpcq==3.7.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
sympy==1.7.1
terminado==0.9.2
testpath==0.4.4
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
z3-solver==4.8.10.0
